pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "pygithub>=2.8.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
]
